requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
cloudinary==1.36.0
Flask-CORS==4.0.0
cryptography==42.0.5
